requirements.txt
odrive>=0.5.4
pyserial>=3.5
numpy>=1.19.3
pandas>=1.1.5
scipy>=1.5.4
matplotlib>=3.3.3
openpyxl>=3.0.5
